Extract CurrentTracker from Player file
